backend/app/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
lightgbm
joblib
yfinance
ta
python-multipart
redis
aioredis
python-jose
passlib
pydantic<2.0.0
python-dotenv
sqlalchemy
fastapi-jwt-auth
pydantic[email]
bcrypt
